Set requirements to something that will yield a work setup in a clean environment.
I was struggling quite a bit to find a working combination of things, and this selection of versions seems to do the trick.

It's restrictive, but perhaps good to keep around in case other people not familiar with these frameworks also try to setup Codespeed. I failed with Django 1.4, and 1.6. From the 1.5 line, other versions might work as well. And mimeparse wasn't a requirement before, but was a missing import.