Better SVN error message.
Error messages reported by Pysvn are usually quite good. They easily allow to
understand the issue and then fix it. Codespeed swallows this nice error message
to replace it by a generic one hiding what was wrong.

For example is anonymous users are not allowed, you will never be able to figure
it out with the "Could not resolve" message.

It seems better to expose the reported error as is rather than trying to "improve"
it.