Add the 'makeimage' service
This adds a service for generating chart images based on the latest
benchmark results using the matplotlib module.
The URL for this service is /sample_project/makeimage/.
The image can be directly embedded in any HTML document using
<img src="host_name/sample_project/makeimage/?...> or simply accessed
using the same URL.

The mandatory GET params are:
env -	the environment's name
proj -	the project's name
branch -the branch's name
exe -	the executable's name
ben -	the benchmark's name
The following params are optional:
revs -	positive integer, number of results the image will display.
	revs=10 is the implicit value representing that the latest 10
	results will be displayed on the chart
relative - if '1' or 'yes', instead of displaying absolute values,
	the first value from the series is used as a baseline and
	the chart will display percentual increments from that baseline
base_commit - specifies a commit name to be used as a baseline instead
	of the first value of the series
base_env - the environment's name to be used for the base_commit instead
	of env
base_proj - the project's name to be used for the base_commit instead
	of proj
base_exe - the executable's name to be use for the base_commit instead of
	exe