Refactor timeline grid to stream data and partition it
The page size is configurable as a setting.
Streaming is used to be able to already send data to the client while more database query my still need to be performed.

timeline.js will keep requesting data as long as the nextBenchmarks field indicates an indicates (!== false).
The plotgrid is only reset on the first request.

Signed-off-by: Stefan Marr <git@stefan-marr.de>